Add appsettings to gitignore
