+++ .gitignore
@@ -38,6 +38,8 @@ bld/
 # Uncomment if you have tasks that create the project's static files in wwwroot
 #wwwroot/
 
+.SignUpApi/appsettings.json
+
 # Visual Studio 2017 auto generated files
 Generated\ Files/
