+++ SignUpApi/appsettings.json
@@ -1,0 +1,50 @@
+{
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft": "Warning",
+      "Microsoft.Hosting.Lifetime": "Information"
+    }
+  },
+  "AllowedHosts": "*",
+
+  "JWT": {
+    "Audience": "http://localhost:4000",
+    "Issuer": "http://localhost:4000",
+    "Key": "StrONGKAutHENTICATIONKEy",
+    "accessTokenExpiration": 20,
+    "refreshTokenExpiration": 60
+  },
+  "AppSettings": {
+    "Secret": "StrONGKAutHENTICATIONKEy"
+  },
+  "MicrosoftAppId": "4c9badde-539f-43ef-9738-3c51893358d7",
+  "MicrosoftAppPassword": "ixv7Q~mBQ5TS5WRTDL5Gpth4j0bcdTbEMqxHb",
+  "AzureAd": {
+    "Instance": "https://login.microsoftonline.com/",
+    "Domain": "cxunicorn.com",
+    "TenantId": "f97015b5-eb92-47be-949e-0ac1c2456f66",
+    "ClientId": "4c9badde-539f-43ef-9738-3c51893358d7",
+    "ClientSecret": "ixv7Q~mBQ5TS5WRTDL5Gpth4j0bcdTbEMqxHb",
+    "Scopes": "access_as_user",
+    "CallbackPath": "/signin-oidc",
+    "SignedOutCallbackPath": "/signout-callback-oidc"
+  },
+  "DownstreamApi": {
+    "BaseUrl": "https://graph.microsoft.com/v1.0",
+    "Scopes": "https://graph.microsoft.com/.default"
+  },
+  "ConnectionStrings": {
+    "WebApiDatabase": "Server=(localdb)\\mssqllocaldb;Database=Assessment;Trusted_Connection=True;"
+  },
+  "SmtpSettings": {
+    "Server": "smtp.gmail.com",
+    "DefaultCredentials": false,
+    "Port": 465,
+    "SenderName": "Rami",
+    "SenderEmail": "rami13195@gmail.com",
+    "UserName": "rami13195@gmail.com",
+    "Password": "acmpform",
+    "SSL": true
+  }
+}
